Separate PySB Preassembler
